requirements.txt
# Core
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
pyyaml>=6.0

# Machine Learning
scikit-learn>=1.0.0
lightgbm>=3.3.0
tensorflow>=2.10.0

# Visualization
plotly>=5.0.0
matplotlib>=3.5.0

# Utilities
python-dateutil>=2.8.0
